fix: extract review_lease + managed_repos submodules to keep binding.rs under 2500 LOC

Agend-Agent: claude-aef7c0
Agend-Task: t-20260709012405047317-61315-3
Agend-Branch: fix/review-branch-lifecycle-settlement
Agend-Issued-At: 2026-07-15T12:32:22.059716+00:00
