+++ src/binding.rs
@@ -454,128 +454,8 @@ pub fn bind_full(
     Ok(())
 }
 
-/// Augment an existing binding with typed review-lease provenance.
-///
-/// Called by the dispatch infrastructure AFTER `bind_full` when a
-/// review-assignment is authority-proven (typed assignment + task + reviewer
-/// target). The fields are added to binding.json under the same locks and
-/// re-signed, so the HMAC covers them — an agent cannot forge these fields.
-///
-/// **Authority chain**: dispatch_auto_bind_lease → review assignment validated
-/// by the daemon → bind_full (creates binding) → this function (annotates with
-/// typed lease_kind + assignment_id + expected_head). The HMAC signature
-/// prevents an agent from adding/modifying these fields itself.
-pub(crate) fn augment_binding_with_lease(
-    home: &Path,
-    agent: &str,
-    lease_kind: &str,
-    review_assignment_id: &str,
-    expected_head: &str,
-) -> Result<(), String> {
-    let _agent_lock = acquire_agent_mutation_lock(home, agent)?;
-    let _binding_lock = acquire_binding_file_lock(home, agent)?;
-    let dir = crate::paths::runtime_dir(home).join(agent);
-    let path = dir.join("binding.json");
-    let content = std::fs::read_to_string(&path)
-        .map_err(|e| format!("read binding for lease augment: {e}"))?;
-    let mut binding: serde_json::Value = serde_json::from_str(&content)
-        .map_err(|e| format!("parse binding for lease augment: {e}"))?;
-    binding["lease_kind"] = serde_json::json!(lease_kind);
-    binding["review_assignment_id"] = serde_json::json!(review_assignment_id);
-    binding["expected_head"] = serde_json::json!(expected_head);
-    let body = serde_json::to_string_pretty(&binding).unwrap_or_default();
-    crate::store::atomic_write(&path, body.as_bytes())
-        .map_err(|e| format!("write binding for lease augment: {e}"))?;
-    match agentic_git_core::integrity_core::sign_binding(home, body.as_bytes()) {
-        Ok(tag) => {
-            if let Err(e) = crate::store::atomic_write(&binding_sig_path(&dir), tag.as_bytes()) {
-                tracing::warn!(%agent, error = %e,
-                    "lease augment sig write failed — shim fails closed (deny) until re-bind");
-            }
-        }
-        Err(e) => tracing::warn!(%agent, error = %e,
-            "lease augment HMAC sign failed — shim fails closed (deny) until re-bind"),
-    }
-    if let Ok(mut map) = binding_index().write() {
-        map.insert(index_key(home, agent), binding);
-    }
-    Ok(())
-}
-
-/// Post-bind hook: if the agent has a receipt-capable typed review assignment
-/// for the given task, augment the binding with HMAC-signed lease provenance.
-///
-/// Authority chain: task_id → task.branch (subject branch) → SCM slug from
-/// source_repo → `assignment_authority::get(repo, subject_branch, agent)` →
-/// verify `is_receipt_capable()` + `reviewed_head` matches checkout tip →
-/// `augment_binding_with_lease`. Never infers from branch prefix or uses
-/// legacy/non-receipt-capable assignments.
-pub(crate) fn try_augment_review_lease(
-    home: &Path,
-    agent: &str,
-    task_id: &str,
-    checkout_branch: &str,
-    source_repo: &Path,
-) {
-    if task_id.is_empty() {
-        return;
-    }
-    let Ok(remote_url) = crate::git_helpers::git_cmd(source_repo, &["remote", "get-url", "origin"])
-    else {
-        return;
-    };
-    let Some(slug) = crate::branch_sweep::extract_github_repo_for_intent(&remote_url) else {
-        return;
-    };
-    let task = match crate::tasks::load_routed(home, task_id) {
-        Ok(rt) => rt.task,
-        Err(_) => return,
-    };
-    let Some(subject_branch) = task.branch.as_deref().filter(|b| !b.is_empty()) else {
-        return;
-    };
-    let Some(assignment) =
-        crate::daemon::assignment_authority::get(home, &slug, subject_branch, agent)
-    else {
-        return;
-    };
-    if !assignment.is_receipt_capable() {
-        return;
-    }
-    if assignment.task_id != task_id {
-        return;
-    }
-    let current_id = crate::fleet::resolve_uuid(home, agent);
-    match (&assignment.target_instance_id, &current_id) {
-        (Some(assign_id), Some(cur_id)) if assign_id == cur_id => {}
-        _ => return,
-    }
-    let Some(reviewed_head) = assignment
-        .reviewed_head
-        .as_deref()
-        .filter(|h| !h.is_empty())
-    else {
-        return;
-    };
-    let Ok(tip) = crate::git_helpers::git_cmd(source_repo, &["rev-parse", checkout_branch]) else {
-        return;
-    };
-    if tip.trim() != reviewed_head {
-        return;
-    }
-    if let Err(e) = augment_binding_with_lease(
-        home,
-        agent,
-        "review",
-        &assignment.assignment_id.to_string(),
-        tip.trim(),
-    ) {
-        tracing::warn!(
-            %agent, %task_id, error = %e,
-            "review lease augmentation failed — review branch preserved on release"
-        );
-    }
-}
+mod review_lease;
+pub(crate) use review_lease::try_augment_review_lease;
 
 /// #2158 GR1: surface an OUT-OF-DISPATCH binding CREATE/CHANGE (no task_id) to the
 /// operator — the realistic accidental-sub-agent / first-bind-hijack vector that
@@ -802,76 +682,11 @@ pub(crate) fn binding_scan_all(home: &Path) -> Vec<(String, serde_json::Value)> 
 /// unwatched PR in an otherwise-unseeded repo would never be discovered (the
 /// #1782 gap). Returns raw paths (slug resolution is the caller's job) to keep
 /// this module free of the git/scm dependency.
-pub fn bound_source_repos(home: &Path) -> Vec<std::path::PathBuf> {
-    let mut repos: Vec<std::path::PathBuf> = Vec::new();
-    for (_, v) in binding_scan_all(home) {
-        if let Some(src) = v["source_repo"].as_str() {
-            let path = std::path::PathBuf::from(src);
-            if !repos.contains(&path) {
-                repos.push(path);
-            }
-        }
-    }
-    repos
-}
-
-/// #2800-lifecycle: union of live bindings + durable managed-repo registry.
-/// Closes the last-binding-released repo-discovery hole: after every binding
-/// for a repo is released, the repo stays discoverable for sweep/janitor.
-pub fn all_managed_repos(home: &Path) -> Vec<std::path::PathBuf> {
-    let mut repos = bound_source_repos(home);
-    for path in read_managed_repo_registry(home) {
-        if !repos.contains(&path) {
-            repos.push(path);
-        }
-    }
-    for repo_str in crate::cleanup_intents::intent_repos(home) {
-        let path = std::path::PathBuf::from(repo_str);
-        if !repos.contains(&path) {
-            repos.push(path);
-        }
-    }
-    repos
-}
-
-fn managed_repo_registry_path(home: &Path) -> std::path::PathBuf {
-    home.join("managed-repos.jsonl")
-}
-
-fn register_managed_repo(home: &Path, source_repo: &str) {
-    if source_repo.is_empty() {
-        return;
-    }
-    let path = managed_repo_registry_path(home);
-    let existing = std::fs::read_to_string(&path).unwrap_or_default();
-    if existing.lines().any(|l| l.trim() == source_repo) {
-        return;
-    }
-    use std::io::Write;
-    if let Ok(mut f) = std::fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(&path)
-    {
-        let _ = writeln!(f, "{source_repo}");
-    }
-}
-
-fn read_managed_repo_registry(home: &Path) -> Vec<std::path::PathBuf> {
-    let path = managed_repo_registry_path(home);
-    std::fs::read_to_string(&path)
-        .unwrap_or_default()
-        .lines()
-        .filter(|l| !l.trim().is_empty())
-        .map(|l| std::path::PathBuf::from(l.trim()))
-        .collect()
-}
-
-// #t-…81457-1: the worktree-occupancy equivalent of `bound_source_repos`
-// lives in `worktree_cleanup` (its only consumer) as
-// `bound_worktree_paths_or_ambiguous` — it needs to distinguish "no
-// binding.json" from "unreadable/corrupt binding.json" for the delete-path
-// fail-closed requirement, unlike this module's existing scan helpers.
+mod managed_repos;
+pub use managed_repos::{all_managed_repos, bound_source_repos};
+pub(crate) use managed_repos::register_managed_repo;
+#[cfg(test)]
+pub(crate) use managed_repos::read_managed_repo_registry;
 
 /// Read the current binding for an agent.
 /// Hot path: returns from in-memory index (read lock). Cold path
